added hotel block widget
